renamed authActions, removed useless browse, replaced double equal in actions
